+++ src/material/standard/standard-material.ts
@@ -73,7 +73,7 @@ export class StandardMaterial extends Material {
   set baseColorTexture(value: StandardMaterialTexture | undefined) {
     if (value !== this._baseColorTexture) {
       this.invalidateShader()
-      if (!value?.transform && value?.frame) {
+      if (!value?.transform && value?.frame && !value?.noFrame) {
         value.transform = TextureTransform.fromTexture(value)
       }
       this._baseColorTexture = value
@@ -88,7 +88,7 @@ export class StandardMaterial extends Material {
   set metallicRoughnessTexture(value: StandardMaterialTexture | undefined) {
     if (value !== this._metallicRoughnessTexture) {
       this.invalidateShader()
-      if (!value?.transform && value?.frame) {
+      if (!value?.transform && value?.frame && !value?.noFrame) {
         value.transform = TextureTransform.fromTexture(value)
       }
       this._metallicRoughnessTexture = value
@@ -103,7 +103,7 @@ export class StandardMaterial extends Material {
   set normalTexture(value: StandardMaterialNormalTexture | undefined) {
     if (value !== this._normalTexture) {
       this.invalidateShader()
-      if (!value?.transform && value?.frame) {
+      if (!value?.transform && value?.frame && !value?.noFrame) {
         value.transform = TextureTransform.fromTexture(value)
       }
       this._normalTexture = value
@@ -118,7 +118,7 @@ export class StandardMaterial extends Material {
   set occlusionTexture(value: StandardMaterialOcclusionTexture | undefined) {
     if (value !== this._occlusionTexture) {
       this.invalidateShader()
-      if (!value?.transform && value?.frame) {
+      if (!value?.transform && value?.frame && !value?.noFrame) {
         value.transform = TextureTransform.fromTexture(value)
       }
       this._occlusionTexture = value
@@ -133,7 +133,7 @@ export class StandardMaterial extends Material {
   set emissiveTexture(value: StandardMaterialTexture | undefined) {
     if (value !== this._emissiveTexture) {
       this.invalidateShader()
-      if (!value?.transform && value?.frame) {
+      if (!value?.transform && value?.frame && !value?.noFrame) {
         value.transform = TextureTransform.fromTexture(value)
       }
       this._emissiveTexture = value