+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- Fixed an issue which caused camera orbit control not to function correctly in Firefox.
+
 ## [1.4.0] - 2022-05-01
 ### Added
 - Added a static function to create a `PickingHitArea` without the need to pass a renderer.