+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Fixed an issue which caused picking interaction to not function correctly when resolution was any other value than 1.
 - Fixed an issue which caused skybox to not render correctly when using PixiJS v6+.
+- Fixed sorting of meshes when using values less than 0.
 
 ## [1.4.1] - 2022-06-14
 ### Fixed